Fixes for bug #3418258.  Cleans up all of the 'unchecked' and 'deprecated' linter warnings.  Several of the redundant casts have been intentionally left for clarity. Set '-serial' and '-cast' linter flags to reduce the number of warnings (it should be safe to ignore serial because the end product is a jar rather than individual class files.  AFAIK, there is no performance penalty for redundant casts in Java).  If anyone objects to the linter flags they can be removed (the linter flags, that is).
git-svn-id: svn://svn.code.sf.net/p/tigervnc/code/trunk@4694 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ java/src/com/tigervnc/vncviewer/DesktopWindow.java
@@ -339,8 +339,8 @@ class DesktopWindow extends JPanel implements
         }
       }
     }
-    scaleWidthRatio = (float)((float)scaledWidth / (float)cc.cp.width);
-    scaleHeightRatio = (float)((float)scaledHeight / (float)cc.cp.height);
+    scaleWidthRatio = (float)scaledWidth / (float)cc.cp.width;
+    scaleHeightRatio = (float)scaledHeight / (float)cc.cp.height;
   }
 
   synchronized public void paintComponent(Graphics g) {
@@ -475,8 +475,8 @@ class DesktopWindow extends JPanel implements
       cursorVisible = true;
       if (softCursor != null) return;
 
-      int cursorLeft = (int)cursor.hotspot.x;
-      int cursorTop = (int)cursor.hotspot.y;
+      int cursorLeft = cursor.hotspot.x;
+      int cursorTop = cursor.hotspot.y;
       int cursorRight = cursorLeft + cursor.width();
       int cursorBottom = cursorTop + cursor.height();
